:truck: Moved config to own package.
